pyproject.toml
[tool.poetry]
name = "automation_indra_avitech"
version = "0.1.0"
description = ""
authors = ["Jan Kamenicky <jankamenicky123@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
pytest = "^8.3.5"
pytest-bdd = "^8.1.0"
pytest-playwright = "^0.7.0"
pytest-html = "^4.1.1"
pyaml = "^25.1.0"
cattrs = "^24.1.3"
cryptography = "^45.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--browser chromium --slowmo 150 --html=\"reports/report.html\" --self-contained-html"

[pytest]
generate_report_on_test = true

